FEATURE: Separating general utils from cyberglove utils to decouple cyberglove dependencies in linux build
